refactor(fixTJunctions): moved fixTJunctions out to a seperate module
